pyproject.toml
[project]
name = "trading-research"
version = "0.1.0"
description = "Personal quant trading research lab"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "tenacity>=8.2",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "pandas-market-calendars>=4.4",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "click>=8.1",
    "typer>=0.12",
    "dash>=2.17",
    "plotly>=5.22",
    "scipy>=1.10",
    "scikit-learn>=1.4",
    "shap>=0.44",
    "lightgbm>=4.3",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5",
]

[project.scripts]
trading-research = "trading_research.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading_research"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q --strict-markers"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "SIM", "RET"]
ignore = ["E501"]
